Specify dtype for supportTensor in ehexFFT.py

+++ EHFFT/ehexFFT.py
@@ -28,7 +28,7 @@ def hexFFT(mat):
   diagDown = torch.zeros(mat.shape,dtype=torch.complex64)
   vert = torch.zeros(mat.shape,dtype=torch.complex64)
 
-  supportTensor = torch.zeros(2*mat.shape[1])
+  supportTensor = torch.zeros(2*mat.shape[1],dtype=torch.complex64)
 
   for i in range(mat.shape[0]//2):
     for j in range(2*mat.shape[1]):
@@ -60,7 +60,7 @@ def hexIFFT(vert, diagUp, diagDown,kernel=[.5,.25,.25]):
   diagDownIFT = torch.zeros(vert.shape,dtype=torch.complex64)
   vertIFT = torch.zeros(vert.shape,dtype=torch.complex64)
 
-  supportTensor = torch.zeros(2*vert.shape[1])
+  supportTensor = torch.zeros(2*vert.shape[1],dtype=torch.complex64)
   for i in range(diagUp.shape[0]//2):
     for j in range(2*diagUp.shape[1]):
       ptsUpX = (i+hexUpDiag(j)[0])%diagUp.shape[0]
